Hi Cheok,
I setup an Computer that is not in the same domain as the proxy and it works also in our network. It was my proposal, i think we can drop the domain field and when somebody this need, we can activate these field again.

Are you happy with this solutions?

best regards
Peter