Bugfix:  Update warning_message_running_2_jstock to clearer message. Bugfix:  Better UI. We will scroll and highlight the newly added stock.
Thanks a lot for Your quick response. I would jist dare to recomed You to change the text of the message, because I didn't understood this meaning at all. It seemed to me as some virus also due to different graphics, and I even uninstalled the JStock.

Thanks and good luck with the SW