Thanks for replying so fast.
1) Yes it only happens when I quit jStock, but not when I use the File > Exit method. It only happens if I click on the actual java window and click Quit org.yccheok.jstock.gui.MainFrame or press command+Q the shortcut for quit.
2) It hangs means that it does not close when I tell it to close/quit and so it just stays open and nothing I click on in the application works, and so I have to force quit the application.
3) this behavior does not happen in 1.0.7h and I have the latest version of java for my specific Mac OS.

Now, as i looked over the code you provided, I believe line 426 calls the image icon file on close, but cannot find it for Mac.

I actually dont remember much of Java programming because I took that class many years ago, but line 2040 calls the Utils for Windows 7 and Windows 8, and not MacOS.
Maybe that might be it? Not sure why it closes perfectly fine when I use the File > Exit but not when command+Q.
Maybe compile it with Utils.isMac() as well? I wouldnt want you to waste time with this, but if its really easy to just add the line for mac it just helps users that are on mac computers and use command+Q to close programs easier.

Thanks!