Typo fix and ChartJDialog mods
Fixed a type in gui_en.properties and made CharJDialog to be JFrame
based so it can easily be maximized. The behaviour otherwise is still
largely the same in terms of window behaviour and focus control. The one
side effect that I cannot remove is that the chart form is now shown on
the Task Bar.